+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to the public API will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Changes to be tagged
+
+## 5.70.0 - 2021-08-14
 ### Added
+- raw `svtScript` endpoint.
 - `mstSvtIndividuality` to raw svt and `traitAdd` to nice svt.
 - charaGraphEx to servant assets.
 - skillAdd to skill.